services/rag/pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
description = "RAG service for Ansible Log Monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.37.0",
    "httpx>=0.27.2",
    "sqlmodel>=0.0.8",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.30.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "minio>=7.2.17",
]

